Add Middleware

+++ app/categories/controller.js
@@ -1,8 +1,17 @@
 const Category = require('./model');
+const {policyFor} = require('../policy');
 
 
 async function store(req,res,next) {
     try {
+      let policy = policyFor(req.user);
+      if(!policy.can('create', 'Category')){
+        return res.json({
+          error: 1,
+          message: `Anda tidak memiliki akses untuk membuat category`
+        });
+      }
+
         let payload = req.body;
         let category = new Category(payload);
         await category.save();
@@ -24,6 +33,13 @@ async function store(req,res,next) {
 async function update(req,res,next) {
     try {
         let payload = req.body;
+        let policy = policyFor(req.user);
+        if(!policy.can('update', 'Category')){
+          return res.json({
+            error: 1,
+            message: `Anda tidak memiliki akses untuk mengubah category`
+          });
+        }
         let category = await Category.findOneAndUpdate({_id: req.params.id}, payload, {new: true, runValidators: true});
 
         return res.json(category);
@@ -41,6 +57,13 @@ async function update(req,res,next) {
 
 async function destroy(req,res,next) {
     try {
+      let policy = policyFor(req.user);
+      if(!policy.can('delete', 'Category')){
+        return res.json({
+          error: 1,
+          message: `Anda tidak memiliki akses untuk menghapus category`
+        });
+      }
         let deleted = await Category.findOneAndDelete({_id: req.params.id});
 
         return res.json(deleted);
